+++ service/.codegen/interface.go.tmpl
@@ -5,7 +5,7 @@ package {{.Name}}
 import (
 	"context"
 	"io"
-	"github.com/databricks/databricks-sdk-go/retries"
+	"github.com/databricks/databricks-sdk-go/databricks/retries"
 )
 
 {{range .Services}}
